worked on getting dtabase to populate on frontend and still does not work
