+++ BookModel.js
@@ -1,15 +1,15 @@
-'use strict';
+// 'use strict';
 
-const mongoose = require('mongoose');
+// const mongoose = require('mongoose');
 
-// this is from the class demo
-const bookSchema = new mongoose.Schema({
-  title: {type: String, required: true},
-  description: {type: String, required: true},
-  status: {type: String},
-  email: {type: String, required: true},
-})
+// // this is from the class demo
+// const bookSchema = new mongoose.Schema({
+//   title: {type: String, required: true},
+//   description: {type: String, required: true},
+//   status: {type: String},
+//   email: {type: String, required: true},
+// })
 
-const BookModel = mongoose.model('books', bookSchema);
+// const BookModel = mongoose.model('books', bookSchema);
 
-module.exports = BookModel;
+// module.exports = BookModel;
