+++ README.md
@@ -22,6 +22,7 @@ This is a React App utilizing a Node JS server that uses Auth0 to exchange a tok
 React, Node JS, Mongo DB, Mongoose, GitHub, Auth0, JWKS, Trello
 
 ## Change Log
+
 <!-- Use this area to document the iterative changes made to your application as each feature is successfully implemented. Use time stamps. Here's an example:
 
 01-01-2001 4:59pm - Application now has a fully-functional express server, with a GET route for the location resource. -->
